large refactoring
this basically reorganize the code in a different logic, closer to
the architecture pattern of elm.

Most files have a model, update and view section now, instead of
separating the code in model/update/view, it is now regrouped by
topic.

Also, a Screens folder was introduced, to collect the top level
components.